Add driver binary for local network
Add a program that runs a local network of gossip nodes
where each node is running on its own threads.

Seems to work, though at some point after some thousands of elements
the gossip stops really working so still need to debug that.

Also this is only implemented for uniform gossip, want to implement the
preferential one as well.