Add preferential gossip to lset
Add support for preferential gossip network to lset. Had to change types
a bit to minimize code duplication between the different types of
gossip.

Seems to work, but in preferential gossip messages always seem to get
stuck? Some debugging in order, but it builds and old functionality
works so committing here as a good checkpoint.