Add loss timeout to lset
Add a timeout to the lset program so that it gives up on an element if
it doesn't appear after that timeout and declares it lost. Makes results
from smaller gossip networks (where a small percentage of messages
are lost) more meaningful rather than just hanging whenever one
message is lost.